(PUP-490) Remove partial import-support from future evaluator
The support is removed (basically a mapping of a function call
to the function "import" to a ImportExpression + removing tests.

The intermediate future parser is still using ImportExpression, so
model and related logic not changed, but marked with deprecation
and "to be removed in Puppet 4" (or whenever the intermediate parser
is removed from the code base).

The grammar/parser is modified, but this is only a removal of a comment.