(PUP-866) Deprecate Import
This deprecates the use of import.
A deprecation warning is generated with a URL to a link that in
turn points to the original Redmine issue.

A test is added.

Note this does not deprecate the underlying "import" mechanism as it is
used for all types of loading of manifests. The deprecation is for
the puppet language construct.