Maint: Standardize docs for refresh behavior in resource types; clarify exec docs
Service/mount/exec already documented that they do the refresh thing, but it
wasn't organized in a predictable way. I'm imposing a convention of a bold
"Refresh" label, a reminder of where events come from, and a description of what
that type does when it gets one.

This commit also adds a lot more detail on how execs respond to refresh, since
it's actually really complicated and not the least bit intuitive. I also reworded
some of the older language near the top of the exec type, while I was in there.