(PUP-865) Add ability to parse a directory of manifests
This adds the ability to run a directory of manifests instead
of just a single file. Simply assign a directory to Puppet[:manifest]
via settings or from the command line.

The files are parsed and combined in alphabetical order as if the
files were to be concatenated into a single file.

The support is in environment, and it is only support for the 
initial import. This also means that both current and future parser
can be used (i.e. the support for this is not in the parser).

Testing is done in integration/node/environment_spec.