(#21641) Windows puppet service should log to the eventlog
Initial attempt to add in eventlogging for all log messages
New service command line argument (--logtofile). By default the service will log all messages to the Eventlog.  This argument will log messages both to the Eventlog and the Puppet log file in Common Application Data.  This can be useful if the EventLog logging is failing or another application depends on the the logfile
Basic tests show the new code works however more testing should be done