(PUP-1048) Windows puppet service should log to the eventlog
Cleaned up some whitespace issues
Changed the wording on some of the log entries to avoid confusion between 'Service resuming' and a Service Resume control message
Added logging for the Pause, Resume and Shutdown service control messages.  Previously there was none
Noted Bug id 11 in the win32-service ruby code
Added a commented out handler for the Interrogation event for debuggin purposes
Changed the name of the event log function from raise_windows_event to report_windows_event to avoid connotations with raise
Changed the event source to Puppet so that the correct event source DLL is used.  However it may be a good idea to split the event source between the service and the actual agent so it's easier to read in the Eventlog.  This would require a change to the puppet_for_the_win repo as well.
The eventlog.close command is within an ensure block to stop leaked handles if an error occurs
The --logtofile argument was working for the service but was causing the puppet agent run to fail as it didn't know what to do with it.  Used a simple text substitution to remove it from the args string.
Spelling mistake
Rebasing broke detecting of the --logtofile argument.  Changed the variable used
Changed wording of the log message when putting the service into a paused state.  Now references a Puppet Issue ticket for tracking
The ensure block will make sure that the eventlog handle is not nil before trying to close the handle