(maint) Move 'jeff_append_to_array' from acceptance to integration test
The test only checked if it was possible to append to an array
in an outer scope by creating two notify resource based on the
array. This test does not have to run as part of the acceptance tests
and was moved to catalog_spec in integration.

This also makes catalog_spec run for both current and future parser.
(The future parser still has issues at this point; to be fixed in
later commits).