(maint) Fix problems with scope bound? and exist? methods
The exist? method did not consider inherited and containing scopes
when checking for existence. The bound? method should have
checked if the variable is bound in the scope used for assignment,
not just the emphemeral-most scope (which may not be assignable).

Fix also guards against ephemeral stack being empty and then
switching to the root @symtable (a corner case that probably never
occurs).