(Maint) Revert including augeas
Including augeas in the tests requires that we have libaugeas installed
on any machine that is going to run the tests with the :extra group. Too
many of our internal systems don't have that kind of setup. It will have
to be up to the maintainer of the augeas code (likely Dominic Cleal) to
make sure that the augeas code works until we can get more complete
testing infrastructure in place.

The ruby-augeas gem can be included from a Gemfile.local for anyone who
wants to develop against augeas in the meantime.