(maint) Update test to work on non-UTF-8 systems
Not all systems run as UTF-8 (our solaris tests run as US-ASCII). In
those cases \xc2\xa0 (NO-BREAK SPACE) is not a valid character. This
causes the test to fail because of encoding problems. By using \x01 (a
control character) we can check the same scenario, but using a character
sequence that is valid in all of our setups.