(maint) Refactor Binder to not use two step creation.
Earlier it was required to first create an instance of Binder,
and then ask it to configure categories, and then layers. Now
that categories are removed, the creation can be made in one
step and the layered bindings are now passed to the constructor.

This removes the need to have methods to query if the binder
is configured.

This also adds a first step in supporting creation of a stacked
injector (the binder can not accept a parent binder - this is w.i.p).