(maint) Acceptance test selector tests for a string rather than a type
Bare upper case words are types in the future parser, so this test
fails because $operatingsystem returns a string which gets compared to
the Resource[Fedora] type.  Explicity quoting so it's clear what the
intention is in the current parser and so that it works properly with
the future parser.