(maint) Simply confine test to any platform with Ruby shadow
Original confine was prone to errors in platform pattern, which we saw
when testing fedora18 whose platform tag no longer contains 'fc-' in our
current acceptance/config/nodes.  This would fall through to the false
case, which meant the host was /not/ rejected.  And then fedora18 from a
source checkout would run the test and fail, because Ruby shadow lib is not
installed by default.  This only turned up because the future parser
acceptance job was running from source checkouts (because one of the
nodes it was testing is Windows...)  At any rate, it proved confusing.

This change just simplifies the confine to reject a host if the Ruby
shadow library is not present.  Because without that, we can't
manipulate passwords, and have nothing to test.