(PUP-672) Puppet::Context is immutable.
Previous implementation exposed a bind method allowing you to add
bindings to the current context. This method has been removed. Now
every context is generated only once, either from an initial push or
override, both of which require a hash of bindings which are merged into
the parent's binding state.