(PUP-1144) Fix up failing tests (pending new/better validation)
Two tests that expected validation to fail did not get triggered
because the too strict validation was removed. It needs to be replaced
with a different better validation. Such validation should be added
in 3.5 (too much work for 3.4) - instead, just accepting potentially
faulty names (they still have to lex correctly).