(PUP-716) Change back to stubs from mocks
During changes a few calls had been inadvertently changed from stubs
(any number of calls allowed, including 0) to mocks (at least one call
required). In the case of the file_set this caused a problem for when
there was no expectaiton of the methods being called. In the daemontools
case, this exposed a mistake that was originally in the tests where the
stub was not even needed.