(PUP-800) Add offset and length to Locatable base class in model
This adds offset and locatable as attributes in the pops model
and makes all objects in a Program model have this trait (except
the top level Program which is always from 0 to the entire length of
the source).

Further, the Program may now contain the line index, and may
contain the source text.