(PUP-1120) Stop pretending other can read CA files
The `cadir` setting controls permissions on the `ca` directory. It
sets permissions to service:service and mode 0770.

    :cadir => {
      :default => "$ssldir/ca",
      :type => :directory,
      :owner => "service",
      :group => "service",
      :mode => 0770,

Previously, the permissions for subdirectories and files made it
appear as though `other` could read/traverse, when in reality they
cannot due to permissions on the `ca` directory.

    # tree -p /var/lib/puppet/ssl
    /var/lib/puppet/ssl
    +-- [drwxrwx---]  ca
    |   +-- [-rw-rw-r--]  ca_crl.pem
    |   +-- [-rw-rw----]  ca_crt.pem
    |   +-- [-rw-rw----]  ca_key.pem
    |   +-- [-rw-r--r--]  inventory.txt
    |   +-- [drwxrwx---]  private
    |   |   +-- [-rw-rw----]  ca.pass
    |   +-- [drwxr-xr-x]  requests
    |   |   +-- [-rw-r-----]  arcturus.local.pem
    |   +-- [-rw-r--r--]  serial
    |   +-- [drwxrwx---]  signed
    |       +-- [-rw-r-----]  puppetmaster.solar.lan.pem

Note `ca_crl.pem` has mode 0664; `inventory.txt` has 0644; `requests`
has 0755, and `serial` has 0644.

This commit cleans up the file permissions so that they are
consistently 0660 for files (where `service` is both owner and group),
and 0770 for the `requests` directory.