(PUP-736) Remove description from RPM/DPKG
Encoding mis-match causes package prefetching to fail with
'invalid byte sequence in US-ASCII'. The guidelines for the
Debian Project [1] and the Fedora Project [2] specify that
metadata files should be UTF-8 encoded, but it is not
enforced.

Since we are not able to force encoding "UTF-8" to work around
this we have decided to back out the description from package
formats for now while working on a longer term solution for
encoding.

Without this fix the package formats fail if there is a package
with UTF-8 encoding in the description.

[1]: http://www.debian.org/doc/debian-policy/ch-controlfields.html
[2]: http://fedoraproject.org/wiki/Packaging:Guidelines#Encoding

Paired with Josh Partlow <joshua.partlow@puppetlabs.com>