(PUP-800) Clean up and improve handling of location / position
There were several issues with the implementation of location
handling - primarily missing the ability to be able to
serialize/de-serialize the model with retained positioning information.

This changes the design to bring the two values offset and length
into each model object and that location information is recored in
the top level Program construct. With this change a Program is
always returned as the result of a parse.

The change touches many files, due to the change to always return
a Model::Program. Most of the change is removal of now no longer
needed constructs.