(PUP-736) Do not test for package not found
And log unmatched lines at debug level. The executors for dpkg and rpm
will raise an error if the command line returns a non-zero status, which
it does for package not found errors.  This bypasses parsing and is
interpretted as package absent result.  Therefore there is no need to
attempt to parse 'package not found' and we can avoid any issues with
changing or translated messages.

If there are unrecognizable lines of output returned from a successful
execution, log them at debug level.  Logging at a warning level has the
potential to produce too much noise.

Dpkg#query is no longer multi-line output and no longer requires
filtering through a pipe.