(PUP-1407) Add authorityKeyIdentifier to CRL to make the CA conform to RFC5280
RFC5280 requires CRLs to contain the authorityKeyIdentifier extension[1].
It identifies the public key that corresponds to the private key used to
sign the CRL.

RFC5280 also requires the use of the key identifier method. This
implementation uses the key identifier format described in OpenSSL[2].
In particular, the `always` value is used to require that the extension
is present.

Also note that we need to invoke the `add_extension` method instead of
overwriting the extensions.

[1] http://tools.ietf.org/html/rfc5280#section-5.2.1
[2] http://www.openssl.org/docs/apps/x509v3_config.html#Authority_Key_Identifier_