(PUP-1151) Have puppet depend on json
In order to have the v2 api emit and consume JSON puppet will need to
have access to a json library. This adds it to our dependencies for
debian, redhat, and gem packages. This should be a safe thing to do
since puppet already needs hiera and hiera depends on JSON, so this is
just taking the same dependency that hiera already had and applying it
directly to puppet.