(PUP-1151) Inject source of environments
This pulls out the source of the environments from the V2::Environments
code and instead injects it from a "provide"d handler. The "provide"
method on V2 allows a handler to be constructed lazily, which is needed
when it needs to be instantiated with information that isn't known until
request time (in this case the source of the environments).