(PUP-1938) Remove Debian service provider default on Ubuntu
As of Ubuntu 8.04 Ubuntu uses Upstart instead of the regular Debian init
scripts, and legacy sysvinit commands have been updated to use Upstart
as well. Because the legacy commands provide compatibility and the
Upstart provider can fall back to Debian init scripts if needed, we no
longer need to set the debian service provider as default on Ubuntu.