(PUP-1470) Make mk_resource_method with false.
Previously the method generated by mk_resource_methods create getters
that would convert any "falsey" values to `:absent`. This has the
unfortunate effect of making `false` turn into `:absent`, rendering
these generated methods difficult to use for boolean parameters.

This change allows boolean values to be preserved. It technically
changes the behavior of the method in an incompatible manner, but the
correctness of the new behavior was deemed to outweigh any problems.